Clean separation fo plugins (see https://github.com/geobretagne/mdchecker/issues/2)
